requirements.txt
appdirs==1.4.4
astroid==2.4.1
attrs==19.3.0
bcrypt==3.1.7
black==19.10b0
blinker==1.4
cffi==1.14.0
click==7.1.2
dnspython==1.16.0
email-validator==1.1.1
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.4.2
Flask-WTF==0.14.3
idna==2.9
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
mypy-extensions==0.4.3
pathspec==0.8.0
Pillow==7.1.2
pycodestyle==2.6.0
pycparser==2.20
pyflakes==2.2.0
pylint-plugin-utils==0.6
regex==2020.5.14
six==1.15.0
SQLAlchemy==1.3.17
toml==0.10.1
typed-ast==1.4.1
typing-extensions==3.7.4.2
Werkzeug==1.0.1
wrapt==1.12.1
WTForms==2.3.1
